requirements.txt
fastapi 
uvicorn
pgvector
openai
python-dotenv
pypdf
openai
python-dotenv
aiofiles
langchain
langchain-openai
faiss-cpu
sentence-transformers
python-dotenv
langchain-community
langchain-core
langchain_text_splitters
langchain-huggingface
psycopg2-binary
boto3
pypdf2
llama-cloud-services 
llama-index-core 
llama-index-readers-file 
python-dotenv
llama-index-embeddings-openai
llama-index-vector-stores-supabase
llama-index
s3fs
llama-index-readers-s3